requirements.txt
django==3.1.7
django-debug-toolbar==3.2
dj-database-url==0.5.0
Pillow==8.1.2
environs==9.3.1
djangorestframework==3.12.2
django-phonenumber-field==5.0.0
pytest-django==4.1.0
